Edited entry to the treadmill in run_treadmill
Movement from the wall (North) reduced by 2 centimeters to better align with the treadmill. Motion to the treadmill (East) reduced by 3 centimeters. Before making this adjustment, my robot would consistently be on the very edge of the rollers and it consistently overshot the treadmill to the East wall. Video available to explain my issue.